+++ backend/src/presentation/controllers/signup/signup.spec.ts
@@ -1,7 +1,5 @@
-import { AccountModel } from '../../domain/models/account'
-import { AddAccount, AddAccountModel } from '../../domain/useCases/add-account'
-import { InvalidParamError, MissingParamError, ServerError } from '../errors'
-import { EmailValidator, BooleanValidator } from '../protocols'
+import { EmailValidator, BooleanValidator, AddAccount, AddAccountModel, AccountModel } from './signup-protocols'
+import { InvalidParamError, MissingParamError, ServerError } from '../../errors'
 import { SignUpController } from './signup'
 
 const makeAddAccountStub = (): AddAccount => {
