+++ backend/src/presentation/protocols/index.ts
@@ -1,4 +1,3 @@
 export * from './controller'
-export * from './email-validator'
 export * from './field-validator'
 export * from './http'
